rename history-pane -> console-pane
